# openapi.yaml
openapi: 3.0.0
info:
  title: 'Delilah Resto '
  description: API for delicious food orders. As a customer you can register, see the list of our products and place an order.Restaurant managers have the ability to receive orders and update them..
  termsOfService: https://github.com/hmfarias/03-acamica-delilah#readme
  contact:
    name: Marcelo Farias
    email: hmfarias7@gmail.com
  version: 1.0.0
servers:

  - description: Local Server
    url: http://localhost:3000/api/v1
...
paths:
...
  /paymethods:
    # ----------------------------------------------POST NEW PAY METHOD
    post:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read      
      summary: "Add a new payment method to Delilah Resto's system"
      description: 'Admin level required'
      operationId: newPayMethod
      requestBody:
        description: 'Required to post a new Payment Method'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payMethodReq'
        required: true
      responses:
        '200':
          description: 'Successful operation - returns the added payment method'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodResp200'      
        '400':
          description: 'Some required fields is not present'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400'
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '409':
          description: 'Payment method already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error409'           
        '500':
          description: 'The payment method could not be registered'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'    
    # --------------------------------------------GET ALL PAYMENT METHODS          
    get:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read      
      summary: 'List all payment methods'
      description: 'Admin level required'
      operationId: getPayMethods
      responses:
        '200':
          description: 'Successful operation - returns an array of payment methods'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodsResp200'          
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'There are no payment methods in the database'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'           
  /paymethods/{id}:
    # --------------------------------------------GET ROLE BY ID
    get:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read
      summary: 'Find payment method by Id - Return a single payment method'
      description: 'Admin level required'
      operationId: getPayMethod
      parameters:
        - name: id
          in: path
          description: 'ID of payment method to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the payment method with the Id specified by parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodResp200'       
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The role not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        '410':
          description: 'The payment method is already deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /paymethods/update/{id}:
    # -------------------------------------------PUT UPDATE PAYMENT METHOD          
    put:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read
      summary: 'Update an existent payment method'
      description: 'Admin level required'
      operationId: updatePayMethod
      parameters:
        - name: id
          in: path
          description: 'ID of payment method that needs to be updated'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Required to update a payment method'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/payMethodReqUpdate'
        required: true
      responses:
        '200':
          description: 'Successful operation - returns the updated payment method'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodResp200'          
        '400':
          description: 'Id parameter or name field are not present'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'Payment method not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'       
        '500':
          description: 'The payment method could not be updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                                   
        'default':
          description: 'Unexpected error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /paymethods/delete/{id}:
      # ---------------------------------------DELETE PAYMENT METHOD BY ID
    delete:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read      
      summary: 'Deletes a payment method by Id'
      description: 'Admin level required'
      operationId: deletePayMethod
      parameters:
        - name: id
          in: path
          description: 'ID of payment method that needs to be deleted'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the deleted payment method'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The payment method not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '410':
          description: 'The payment method is already deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The payment method could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'    
  /paymethods/restore/{id}:
    # ----------------------------------------RESTORE PAYMENT METHOD BY ID
    put:
      tags:
        - PayMethods
      security: 
        - bearerAuth: 
          - read      
      summary: 'Restore a payment method by Id'
      description: 'Admin Level required'
      operationId: restorePayMethod
      parameters:
        - name: id
          in: path
          description: 'ID of payment method that needs to be restored'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the restored payment method'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/payMethodResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The payment method not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '406':
          description: 'The payment method is not deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error406'                     
        '410':
          description: 'The payment method is deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The payment method could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'        

         
                
#  -----------------------------------------------------------------COMPONENTS
components:
...
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT

  #  --------------------------------------------------------SCHEMAS       
  schemas:
...
    payMethod:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 7
        name:
          type: string
          example: 'user'
        createdAt:
          type: string
          example: '|creation date and time|'
        updatedAt:
          type: string
          example: '|update date and time|'
        deletedAt:
          type: string
          example: '|null| or |delete date and time|'
...
    payMethods:
      type: array
      items: 
        $ref: '#/components/schemas/payMethod'          
    payMethodReq:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: 'user'
    payMethodReqUpdate:
      type: object
      properties:
        name:
          type: string
          example: 'user'
    payMethodResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            payMethod:
              $ref: '#/components/schemas/payMethod'      
        message:
          $ref: '#/components/schemas/messageSuccess'
    payMethodsResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            payMethods:
              $ref: '#/components/schemas/payMethods'     
        message:
          type: string
          example: 'Successfully recovered Payment Methods'

          
    # -----------------------------------SCHEMA FOR Success message      
    messageSuccess:
      type: string
      example: 'Successful operation for |entity| ID: |entity ID|, Name: |entity name|'     
          
    #  ----------------------------------SCHEMAS FOR ERRORS
    errorDefault:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
          properties:
            error:
              type: string
              example: '|Error|'
        message:
          type: string
          example: 'Internal error - Try again later'
    error400:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|field| field is required |'  
    error400Param:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|parameter or field| |parameter or field name| is required | or has an incorrect format'  
    error403:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Admin level required'  
...
    error404:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| not found'     
    error406:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is not deleted'     
    error409:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| |entity name| already exists'     
    error410:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is deleted - (soft deleted)'
    error500:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Unexpected error - |message|'